Fix imports for docs build
